backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
python-dotenv==1.1.1
httpx==0.28.1
requests==2.32.5
browser-cookie3==0.19.1
websocket-client==1.8.0
